feat: use tippy in getting started card instead of fixed position
