--- src/Shared/Components/GettingStartedCard/gettingStarted.scss
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2024. Devtron Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-.getting-started-card {
-    position: fixed;
-    z-index: 6;
-    right: 20px;
-    top: 54px;
-}
-
-.arrow-up {
-    width: 0;
-    height: 0;
-    border-left: 12px solid transparent;
-    border-right: 12px solid transparent;
-    border-bottom: 12px solid var(--bg-overlay-primary);
-    right: 64px;
-    position: fixed;
-    top: 42px;
-}
-
-.getting_tippy__position {
-    position: fixed;
-}
