+++ src/Shared/Components/GettingStartedCard/types.ts
@@ -27,8 +27,5 @@ export interface LoginCountType extends ResponseType {
 }
 
 export interface GettingStartedType {
-    className: string
-    showHelpCard: boolean
     hideGettingStartedCard: (count?: string) => void
-    loginCount: number
 }
